# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Grossess Movies and its Director",
    "description": "A app for the most grosses movies and their directors ",
    "version": "1.0.0"
  },
  "securityDefinitions": {
    "oauth": {
        "type": "oauth2",
        "authorizationUrl": "http://localhost:3000/api-docs",
        "flows": 
        "implicit",
        "scopes": {
            "read:movies": "read movie data",
            "write:movies": "whrite movie data",

            "read:directors": "read director data",
            "write:directors": "write director data"
        }
    }
},
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/movies/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
        
      },

      "post": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rank": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "director": {
                  "example": "any"
                },
                "yearReleaser": {
                  "example": "any"
                },
                "rated": {
                  "example": "any"
                },
                "budget": {
                  "example": "any"
                },
                "lifeTimeGross": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/directors/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                },
                "nationality": {
                  "example": "any"
                },
                "knownFor": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}